Define bathymetry style for basemap

Defines bathymetry style to be used in basemap

Usage

define_bathy_style(x)

Arguments

x

Character argument giving the input bathymetry style. Partially matched and can be abbreviated. See bathy.style in basemap.

Value

Returns a named character vector with the name pointing to the bathymetry style and value to internal map element command for basemap (see map_cmd).

Author

Mikko Vihtakari