Open Geonorge bathymetry shapefiles

Opens and formats Geonorge bathymetry shapefiles ready for plotting in ggOceanMaps

Usage

geonorge_bathymetry(filepath, layer = NULL, verbose = FALSE)

Arguments

filepath

Character string defining the path to the .gml file. Must contain the file extension.

layer

Character string defining the layer containing depth information. If NULL assumed to be "dybdeareal".

verbose

Logical indicating whether information the reading process should be returned.

Value

An sf object containing the depth polygons. Uses same projection than bathy (see CRS).

Details

You can download the bathymetry polygon shapefiles from Geonorge. Download the file in GLM format.

See also

Author

Mikko Vihtakari