Plot Data on Oceanographic Maps using ggplot2

Uses ggplot2 syntax and shape files to plot research data on oceanographic maps anywhere around the globe.

Details

The general map-making function for ggOceanMaps is basemap. This function creates a "canvas" on which research data can be plotted. The basemap function is analogous to ggplot function from the ggplot2 package. Remember to use data = <<NAMEOFDATASET>> for additional geometries you plot on basemaps (basemap(60) + geom_point(data = data.frame(lon = 50, tat = 70), aes(x = lon, y = lat)) as an example).

See also

Author

Maintainer: Mikko Vihtakari mikko.vihtakari@hi.no (ORCID) (Institute of Marine Research)