Return map elements for basemap

An internal function to make basemap code more readable

Usage

map_cmd(command, alternative = FALSE)

Arguments

command

basemap layer to be added

alternative

logical to return alternative formmatting in certain cases. Used to reduce if-else statements in basemap.

Value

A character string containing a ggplot2 plotting command. Use eval(parse(text=...)) to plot the string.

Details

This is an internal function, which is automatically run by the basemap function. Common users do not need to worry about these details. Basemap elements can added together using this function, parse and eval.

See also

Author

Mikko Vihtakari

Examples

## An example for utm map without glaciers or bathymetry
if (FALSE) eval(parse(text=paste(map_cmd("base"), map_cmd("land_utm"),
map_cmd("grid_utm"), map_cmd("defs_utm"), sep = "+"))) # \dontrun{}