Simplify a bathymetry raster ready for vectorization

Simplifies bathymetry raster ready for the vector_bathymetry function. Warning: processing may take a long time if the bathymetry raster is large.

Usage

raster_bathymetry(
  bathy,
  depths,
  proj.out = NULL,
  proj.bathy = NULL,
  boundary = NULL,
  warp = FALSE,
  estimate.land = FALSE,
  downsample = NULL,
  verbose = TRUE
)

Arguments

bathy

A stars object or a string giving the path to a bathymetry NetCDF or grd file

depths

Numeric vector giving the cut points for depth contours (see cut). If NULL, no depth aggregation will be made. This option is suitable for raster bathymetries passed directly to basemap.

proj.out

A character string specifying the coordinate reference system (CRS) argument for the output. See st_crs and proj.org. If NULL, the projection is retrieved from bathy and the output will not be reprojected saving processing time (since proj.out and proj.bathy would match.

proj.bathy

A character string specifying the CRS for the input (bathy). Only required if bathy lacks CRS information. If NULL, "EPSG:4326" is assumed.

boundary

A st_polygon object, text string defining the file path to a spatial polygon, bounding box, or a numeric vector of length 4 giving the boundaries for which bathy should be cut to. Should be given as decimal degrees. If unnamed numeric vector, the first element defines the minimum longitude, the second element the maximum longitude, the third element the minimum latitude and the fourth element the maximum latitude of the bounding box. You can also use the sf bounding box format as named vector. Use NULL not to cut bathy.

warp

Logical indicating whether the resulting grid should be resampled to a new CRS if proj.out != proj.bathy using the st_warp function. A time-consuming operation, but necessary when CRS changes in raster bathymetries. Not required if the next step is to vectorise the bathymetry.

estimate.land

Logical indicating whether to include land to the output. Can be used in the following vector_bathymetry step to estimate land polygons.

downsample

An integer defining how many rows in bathy should be skipped to reduce the size (and resolution). 1 skips every second row, 2 every second and third. See st_downsample. Set to NULL (default) to skip downsampling.

verbose

Logical indicating whether information about progress and guessed projection should be returned. Set to FALSE to make the function silent.

Value

A list with a stars object the containing projected bathymetry defined by the proj.out argument and a data frame of depth intervals.

Details

You can use GEBCO, IBCAO, ETOPO bathymetry grids downloaded from respective sources as the bathy argument. The bathymetry grids read from files must be in any format read by read_stars. Alternatively use the marmap::getNOAA.bathy function to download ETOPO1 bathymetry and convert it to a raster object using the marmap::as.raster function.

Note that the size of the output is heavily influenced by the number of depth contours (depths) as well as the resolution of bathy and choice of downsample. To make the vector_bathymetry function and consequent plotting faster, limiting the details of the bathymetry raster may be desirable.

References

GEBCO Compilation Group (2025) GEBCO 2025 (doi:10.5285/37c52e96-24ea-67ce-e063-7086abc05f29). URL: https://www.gebco.net/data-products-gridded-bathymetry-data/gebco2025-grid. NOAA National Centers for Environmental Information. 2022: ETOPO 2022 15 Arc-Second Global Relief Model. NOAA National Centers for Environmental Information. doi:10.25921/fd45-gt74 .

See also

Author

Mikko Vihtakari