Automatic limits for basemap

Find limits for a basemap from a data frame.

auto_limits(
  data,
  lon = NULL,
  lat = NULL,
  proj.in = "+init=epsg:4326",
  proj.out = NULL,
  expand.factor = NULL,
  verbose = FALSE
)

Arguments

data
Data frame containing data for which the limits should be calculated.

lon, lat
Names of longitude and latitude columns in data as character or integer index. If NULL, the column names are guessed.

proj.in
Original proj4string projection. Must be defined as character argument.

proj.out
Resulting map projection. See transform_coord.

expand.factor
Expansion factor for map limits. Set to NULL to ignore.

verbose
if TRUE, the function prints information about the changed projection. Switch to FALSE to make the function silent.

rotate
Logical indicating whether the limits should be rotated to point towards the pole relative to mid-longitude limit.

Value

A list of limits and projections in proj.in and proj.out formats.

Details

This is an internal function, which is automatically run by the basemap function. The function does some funky illogical action (seeing the argument names) when feeding in projected coordinates. This is to make the automatic limits to work for projected coordinates too. Despite the funkiness, the function may be useful when customising basemaps.

See also

Examples

auto_limits(data = expand.grid(lon = c(-120, 180, 120), lat = c(60, 60, 80)))

#> $ddLimits
#> [1] -100.68400  100.68400   50.94316   63.43154
#> 
#> $projLimits
#> [1] -2886578.7  2886578.7   544589.7  3333134.0
#> 
#> $projBound
#> class       : SpatialPolygons 
#> features    : 1 
#> extent      : -2886579, 2886579, 544589.7, 3333134  (xmin, xmax, ymin, ymax)
#> crs         : +init=epsg:3995 +proj=stere +lat_0=90 +lat_ts=71 +lon_0=0 +k=1 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 
#> 
#> $proj.in
#> [1] "+init=epsg:4326"
#> 
#> $proj.out
#> [1] "+init=epsg:3995"
#>