A list of pre-made shapefiles for basemap

Lists available pre-made shapefiles for plotting in the basemap function. Gives also instructions how to make custom ones.

shapefile_list(name, get.data = FALSE)

Arguments

name
A character argument giving the name of a pre-made shapefile. Will be partially matched. Use "all" to list all available ones.

get.data
Logical indicating whether spatial data should be returned instead of names of spatial data objects.

Value

Returns a data frame of provided pre-made shapefiles, if name = "all". Returns a shapefile list containing the information for a particular map otherwise.

Details

Custom shapefiles for basemap should be defined as lists with (at least) following names (everything should be provided as characters):

land Object name of the SpatialPolygonsDataFrame containing land. Required.

glacier Object name of the SpatialPolygonsDataFrame containing glaciers. Use NULL if glaciers are not needed.

bathy Object name of the SpatialPolygonsDataFrame containing bathymetry contours. Use NULL if bathymetry is not needed.

All linked spatial data objects must be in same projection. Pre-made shapefiles contain additional elements that are used in the basemap function, but not required for custom shapefile datasets.

See also

Author

Mikko Vihtakari

Examples

shapefile_list("all")

#>                     name                            land
#> 1    ArcticStereographic    ggOceanMapsData::arctic_land
#> 2 AntarcticStereographic ggOceanMapsData::antarctic_land
#> 3          DecimalDegree        ggOceanMapsData::dd_land
#>                              glacier                            bathy  crs
#> 1    ggOceanMapsData::arctic_glacier    ggOceanMapsData::arctic_bathy 3995
#> 2 ggOceanMapsData::antarctic_glacier ggOceanMapsData::antarctic_bathy 3031
#> 3        ggOceanMapsData::dd_glacier        ggOceanMapsData::dd_bathy 4326

shapefile_list("Arctic") # partial matching

#> $name
#> [1] "ArcticStereographic"
#> 
#> $land
#> [1] "ggOceanMapsData::arctic_land"
#> 
#> $glacier
#> [1] "ggOceanMapsData::arctic_glacier"
#> 
#> $bathy
#> [1] "ggOceanMapsData::arctic_bathy"
#> 
#> $crs
#> [1] 3995
#>