Create a polygon bathymetry from a raster bathymetry file

Vectorizes bathymetry rasters. Designed to be used for the output of raster_bathymetry function. Warning: processing may take a long time if the bathymetry raster is large.

vector_bathymetry(
  bathy,
  drop.crumbs = NULL,
  remove.holes = NULL,
  smooth = FALSE,
  output.sf = FALSE
)

Arguments

bathy
bathyRaster object from the raster_bathymetry function.

drop.crumbs
Single numeric value specifying a threshold (area in km2) for disconnected polygons which should be removed. Set to NULL to bypass the removal. Uses the drop_crumbs function.

remove.holes
Single numeric value specifying a threshold (area in km2) for holes which should be removed. Set to NULL to bypass the removal. Uses the fill_holes function.

smooth
Logical indicating whether the pixelated contours should be smoothed. Uses the smooth_ksmooth function.

output.sf
Logical indicating whether an sf (TRUE) or sp (FALSE) polygon should be returned.

Value

An sf or sp object containing the depth polygons. Uses same projection than bathy (see CRS).

Details

The drop.crumbs and remove.holes arguments can be used to make the resulting object smaller in file size. The smooth argument can be used to remove the pixelated contours, but often increases file size. Note also that using this option will bias the contours with respect to real world.

See also

Author

Mikko Vihtakari